Fix key-value caching for seqlen != 1
This commit fixes a bug in the key-value caching. Currently,
a square attention mask is misapplied to the scores matrix
despite not matching the shape of the scores matrix. This
results in a runtime error. In a correct implementation, the
decoder mask needs to describe how the new seq_len tokens
interact with all the cached tokens. That is, the attention
mask needs to be of shape (seq_len, total_len), indicating how
the token at row i (representing token i + cached_len in the
transformer model) attends to token j. Accordingly, the matrix
needs to mask entries where j > cached_len + i. This patch
horizontally appends (seq_len, cached_len) zeros to an
upper-triangular mask of size (seq_len, seq_len) to form the
(seq_len, total_len) mask.